Completed Section 4.6 and Homework 4
This section contains a significant update showcasing my work on the final part of the lab, along with the code from the last homework assignment. I’m retaining the homework code because it serves as a valuable reference for correcting FITS images, which will be useful for future labs. NOTE: I need to add corrections to all of the sections that I worked on before submitting my lab report. Please wait for those changes.